+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Added `cache` parameter to `app_runner` to reuse the cache object.
+
 
 ## [1.5.3] - 2022-09-05